+++ previous-versions/accounting/src/codataccounting/bank_account_transactions.py
@@ -161,8 +161,8 @@ class BankAccountTransactions:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[shared.BankTransactionsResponse])
-                res.bank_transactions_response = out
+                out = utils.unmarshal_json(http_res.text, Optional[shared.BankTransactions])
+                res.bank_transactions = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [400, 401, 404, 409, 429]: