+++ previous-versions/accounting/RELEASES.md
@@ -609,3 +609,13 @@ Based on:
 - [python v0.1.0] previous-versions/accounting
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-accounting/0.1.0 - previous-versions/accounting
+
+## 2023-09-05 00:10:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.1] previous-versions/accounting
+### Releases
+- [PyPI v0.1.1] https://pypi.org/project/codat-accounting/0.1.1 - previous-versions/accounting